# app/server/static/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Search Intelligence API",
    "description": "AI-powered career intelligence platform — job discovery, company analysis, skill management, resume generation, and career insights.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Development server" }
  ],
...
  "paths": {
...
    "/api/career-intelligence/{section}": {
      "get": {
        "tags": ["Insights"],
        "summary": "Get latest intelligence for a section",
        "parameters": [
          { "name": "section", "in": "path", "required": true, "schema": { "type": "string", "enum": ["overview", "opportunities", "companies", "market", "networking", "skills_intel"] } }
        ],
        "responses": { "200": { "description": "Section data" }, "404": { "description": "No data found" } }
      }
    },
    "/api/career-intelligence/refresh": {
      "post": {
        "tags": ["Insights"],
        "summary": "Generate all insight sections in background",
        "responses": { "200": { "description": "Generation started" }, "409": { "description": "Already running" } }
      }
    },
    "/api/career-intelligence/{section}/refresh": {
      "post": {
        "tags": ["Insights"],
        "summary": "Generate a single insight section",
        "parameters": [
          { "name": "section", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Generation started" } }
      }
    },
    "/api/career-intelligence/skills-intel": {
      "get": {
        "tags": ["Skills"],
        "summary": "Get latest Skills Intelligence Report",
        "responses": { "200": { "description": "Skills intelligence data" }, "404": { "description": "No data found" } }
      }
    },
...
    "/api/career-intelligence/runs": {
      "get": {
        "tags": ["Insights"],
        "summary": "Get generation history with pagination",
        "parameters": [
          { "name": "section", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "Paginated generation runs" } }
      }
    },
...
  }
}